+++ R/secrets.R
@@ -22,7 +22,8 @@
 set_secret <- function(name, secret = NULL) {
   rlang::check_installed("keyring", "to manage secrets in getRad")
   if (!rlang::is_scalar_character(name)) {
-    cli::cli_abort("{.arg name} should be a scalar character",
+    cli::cli_abort(
+      "{.arg name} must be a single character value.",
       class = "getRad_error_set_secret_no_scalar_character"
     )
   }
@@ -34,11 +35,14 @@ set_secret <- function(name, secret = NULL) {
     )
   }
   sname <- paste0(
-    getOption("getRad.key_prefix",
-      default = cli::cli_abort("The option `getRad.key_prefix` is not found",
+    getOption(
+      "getRad.key_prefix",
+      default = cli::cli_abort(
+        "Can't find the option {.val getRad.key_prefix}.",
         class = "getRad_error_key_prefix_not_found_setting"
       )
-    ), name
+    ),
+    name
   )
   keyring::key_set_with_value(service = sname, password = secret)
   invisible(TRUE)
@@ -60,22 +64,26 @@ list_secrets <- list(
 get_secret <- function(name) {
   rlang::check_installed("keyring", "to manage secrets in getRad")
   if (!rlang::is_scalar_character(name)) {
-    cli::cli_abort("{.arg name} should be a scalar character",
+    cli::cli_abort(
+      "{.arg name} must be a single character value.",
       class = "getRad_error_get_secret_no_scalar_character"
     )
   }
   sname <- paste0(
-    getOption("getRad.key_prefix",
-      default = cli::cli_abort("The option `getRad.key_prefix` is not found",
+    getOption(
+      "getRad.key_prefix",
+      default = cli::cli_abort(
+        "Can't find the option {.val getRad.key_prefix}.",
         class = "getRad_error_key_prefix_not_found_getting"
       )
-    ), name
+    ),
+    name
   )
   if (!(sname %in% keyring::key_list(sname)$service)) {
     cli::cli_abort(
       c(
-        x = "The {.arg {sname}} secret is not in the keyring.",
-        i = 'Please use {.code set_secret("{name}")} to store the secret.'
+        "Can't find secret {.arg {sname}} in the keyring.",
+        "i" = "Please use {.code set_secret(\"{name}\")} to store the secret. Note that the prefix is automatically added in {.fun set_secret}."
       ),
       class = "getRad_error_secret_not_found"
     )