+++ tests/testthat/test-get_pvol.R
@@ -26,7 +26,8 @@ test_that("get_pvol radar argument", {
 })
 
 test_that("get_pvol time argument", {
-  expect_error(get_pvol("nlhrw", datetime = "asdf"),
+  expect_error(
+    get_pvol("nlhrw", datetime = "asdf"),
     class = "getRad_error_time_not_correct"
   )
   expect_error(
@@ -116,7 +117,10 @@ test_that("multiple timestamps and radars work", {
   expect_true(all(unlist(lapply(pvl, bioRad::is.pvol))))
   expect_identical(
     lapply(pvl, \(x) x$datetime),
-    as.list(rep(seq(min(multiple_timestamps), max(multiple_timestamps), "5 mins"), 2))
+    as.list(rep(
+      seq(min(multiple_timestamps), max(multiple_timestamps), "5 mins"),
+      2
+    ))
   )
   expect_identical(
     lapply(pvl, \(x) x$radar),
@@ -126,19 +130,38 @@ test_that("multiple timestamps and radars work", {
 
 test_that("Mixed radar vector (single timestamp)", {
   skip_if_offline()
-  time_utc <- lubridate::floor_date(Sys.time() - lubridate::hours(12), "5 mins")
-  suppressMessages(pvols <- getRad::get_pvol(c("KABR", "czska"), time_utc))
+  time_utc <- lubridate::as_datetime("2021-01-20 05:01:00")
+  suppressMessages(pvols <- getRad::get_pvol(c("KABR", "finur"), time_utc))
   expect_true(is.list(pvols))
-  expect_gt(length(pvols), 0)
+  expect_length(pvols, 2)
   expect_true(all(purrr::map_lgl(pvols, ~ inherits(.x, "pvol"))))
+  expect_equal(
+    purrr::map_chr(pvols, ~ .x$radar),
+    c("KABR", "finur"),
+    ignore_attr = TRUE
+  )
+  expect_identical(
+    pvols[[1]]$datetime,
+    lubridate::as_datetime("2021-01-20 04:57:36 UTC")
+  )
+  expect_identical(
+    pvols[[2]]$datetime,
+    lubridate::floor_date(time_utc, "5 mins")
+  )
 })
 
 test_that("Mixed radar vector + 9 minute interval", {
   skip_if_offline()
-  time_utc <- lubridate::floor_date(Sys.time() - lubridate::hours(12), "5 mins")
+  time_utc <- lubridate::as_datetime("2025-01-20 03:55:50")
   dt_int <- lubridate::interval(time_utc, time_utc + lubridate::minutes(9))
-  suppressMessages(pvols <- getRad::get_pvol(c("KABR", "czska"), dt_int))
+  suppressMessages(pvols <- getRad::get_pvol(c("KABR", "fikan"), dt_int))
   expect_type(pvols, "list")
-  expect_gt(length(pvols), 2)
-    purrr::walk(pvols, ~expect_s3_class(.x, "pvol"))
+  expect_length(pvols, 3)
+  purrr::walk(pvols, ~ expect_s3_class(.x, "pvol"))
+  expect_equal(
+    purrr::map_chr(pvols, ~ .x$radar),
+    c("KABR", "KABR", "fikan"),
+    ignore_attr = T
+  )
+  expect_true(all(purrr::map_vec(pvols, ~ .x$datetime) %within% dt_int))
 })