+++ tests/testthat/test-get_pvol_ro.R
@@ -1,32 +1,26 @@
 test_that("Check if the available attributes changed", {
   skip_if_offline()
   expect_identical(
-    httr2::request(
-      "http://opendata.chmi.cz/meteorology/weather/radar/sites/ska"
-    ) |>
+    httr2::request("https://opendata.meteoromania.ro/radar/MED/") |>
       httr2::req_perform() |>
       httr2::resp_body_html() |>
       xml2::xml_find_all("//a/@href") |>
       xml2::xml_text() |>
-      utils::tail(-1),
-    c(
-      "vol_phidp/",
-      "vol_rhohv/",
-      "vol_u/",
-      "vol_v/",
-      "vol_w/",
-      "vol_z/",
-      "vol_zdr/"
-    )
+      tail(-1) |>
+      gsub(pattern = "MED_[0-9]*00", replacement = "") |>
+      unique() |>
+      gsub(pattern = ".hdf", replacement = "") |>
+      sort(),
+    c("KDP", "RhoHV", "V", "ZDR", "dBZ") |> sort()
   )
 })
-test_that("Pvol for Czechia can be downloaded", {
+test_that("Pvol for Romania can be downloaded", {
   skip_if_offline()
   time <- lubridate::floor_date(
     as.POSIXct(Sys.time(), tz = "Europe/Helsinki") - lubridate::hours(10),
     "5 mins"
   )
-  pvol <- expect_s3_class(get_pvol("czska", time, param = "all"), "pvol")
+  pvol <- expect_s3_class(get_pvol("romed", time, param = "all"), "pvol")
   expect_true(bioRad::is.pvol(pvol))
   expect_identical(
     lubridate::floor_date(pvol$datetime, "5 mins"),